Refatoraçao do modulo util
