+++ vmm_manager/util/config.py
@@ -1,0 +1,16 @@
+"""
+Módulo com configurações gerais
+"""
+
+# Campos customizado no VMM
+CAMPO_AGRUPAMENTO = ('VMM_MANAGER_AGRUPAMENTO',
+                     'Agrupamento da máquina (script vmm_manager).')
+
+CAMPO_ID = ('VMM_MANAGER_ID',
+            'Nome da VM informada no inventário (script vmm_manager).')
+
+CAMPO_IMAGEM = ('VMM_MANAGER_IMAGEM',
+                'Imagem utilizada para criar a VM (script vmm_manager).')
+
+CAMPO_REGIAO = ('VMM_MANAGER_REGIAO',
+                'Região na qual a VM foi criada (script vmm_manager).')
